refactor: add image folder and split file
