+++ client/src/component/atom/Image/index.ts
@@ -1,0 +1,18 @@
+export { default } from './Image'
+
+export namespace ImageType {
+  export interface StyleAttributes extends Object {
+    width?: string
+    height?: string
+    margin?: string
+    padding?: string
+    radius?: string
+    cursor?: string
+  }
+
+  export interface Props {
+    customStyle?: StyleAttributes
+    url?: string
+    onClick?: () => void
+  }
+}
